Create condition_variable_example.txt
This code demonstrates the use of C++ condition_variable for the purpose of event handling in multithreaded environment.
ues case: FBI wants to rescue a child captured by the terrorists. FBI sends one agent to save the child and four soldiers to destroy the terrorist location.
Explanation:
soldiers cover the area and waits for the signal from the agent. once the child is saved, agent gives turns the signal on. upon seeing the signal, soldiers starts attacking.